refactor: add discord manager
