+++ MHFZ_Overlay/Core/Class/Application/ApplicationManager.cs
@@ -1,4 +1,5 @@
 ﻿using DiscordRPC;
+using MHFZ_Overlay.Core.Class.Discord;
 using NLog;
 using System;
 using System.Collections.Generic;
@@ -18,7 +19,7 @@ namespace MHFZ_Overlay.Core.Class.Application
         {
             //https://stackoverflow.com/a/9050477/18859245
             databaseManager.StoreSessionTime(databaseManager.DatabaseStartTime);
-            ApplicationManager.DiscordRPCCleanup();
+            DiscordManager.DiscordRPCCleanup();
             DisposeNotifyIcon();
             logger.Info("Closing overlay");
             Environment.Exit(0);
@@ -35,23 +36,11 @@ namespace MHFZ_Overlay.Core.Class.Application
         public static void HandleRestart()
         {
             databaseManager.StoreSessionTime(databaseManager.DatabaseStartTime);
+            DiscordManager.DiscordRPCCleanup();
             DisposeNotifyIcon();
             logger.Info("Restarting overlay");
             System.Windows.Forms.Application.Restart();
             System.Windows.Application.Current.Shutdown();
-        }
-
-        //Dispose client        
-        /// <summary>
-        /// Cleanups the Discord RPC instance.
-        /// </summary>
-        public static void DiscordRPCCleanup()
-        {
-            if (MainWindow.discordRPCClient != null)//&& ShowDiscordRPC)
-            {
-                MainWindow.discordRPCClient.Dispose();
-                logger.Info("Disposed Discord RPC");
-            }
         }
     }
 }
